Resolve "Update golangci-lint config file to 2.0 format"

Closes #460

See merge request apps/gateway/gateway!418

+++ pkg/pipeline/errors_test.go
@@ -277,17 +277,17 @@ func newTestPipeline(c C, db *database.DB, trans *model.Transfer) *testPipeline 
 	testPip := &testPipeline{Pipeline: pip, transDone: make(chan bool)}
 
 	pip.Trace = Trace{
-		OnPreTask: func(rank int8) error {
+		OnPreTask: func(rank int) error {
 			atomic.AddUint32(&testPip.preTasks, 1)
 
 			return nil
 		},
-		OnPostTask: func(rank int8) error {
+		OnPostTask: func(rank int) error {
 			atomic.AddUint32(&testPip.postTasks, 1)
 
 			return nil
 		},
-		OnErrorTask: func(rank int8) { atomic.AddUint32(&testPip.errTasks, 1) },
+		OnErrorTask: func(rank int) { atomic.AddUint32(&testPip.errTasks, 1) },
 		OnTransferEnd: func() {
 			close(testPip.transDone)
 		},
